Clean up redundant code in VERITAS codebase
- Remove duplicate file header in doctor.py (lines 22-41 were duplicated)
- Remove duplicate imports in pipeline.py, code_planner.py
- Consolidate _safe_float() into new core/utils.py (removed from 6 files)
- Remove duplicate benchmark script run_benchmarks_improved.py
- Remove unused imports (Dict, List, Union) from evolver.py, server.py, atomic_io.py
- Remove commented-out alternative code in code_planner.py
