+++ veritas_os/api/evolver.py
@@ -1,4 +1,4 @@
-from typing import Any, Dict, List
+from typing import Any, List
 import re
 from pathlib import Path
 import json
